moved example files to separate folder
