+++ examples/test2.js
@@ -1,4 +1,4 @@
-var PixelPusher = require('./pixelpusher');
+var PixelPusher = require('../pixelpusher');
 var PixelStrip = PixelPusher.PixelStrip;
 
 new PixelPusher().on('discover', function(controller) {
@@ -23,6 +23,7 @@ new PixelPusher().on('discover', function(controller) {
         console.log('TIMEOUT : PixelPusher at address [' + controller.params.ipAddress + '] with MAC (' + controller.params.macAddress + ') has timed out. Awaiting re-discovery....');
         if (!!timer) clearInterval(timer);
     });
+
     // aquire the number of strips that the controller has said it
     // has connected via the pixel.rc config file
     var NUM_STRIPS = controller.params.pixelpusher.numberStrips;
@@ -34,48 +35,50 @@ new PixelPusher().on('discover', function(controller) {
     var PIXELS_PER_STRIP = controller.params.pixelpusher.pixelsPerStrip;
 
     // create a loop that will send commands to the PP to update the strip
-    var UPDATE_FREQUENCY_MILLIS = 30; // 15 is just faster than 60 FPS
+    var UPDATE_FREQUENCY_MILLIS = 30;// 15 is just faster than 60 FPS
+
+    var waveHeight = PIXELS_PER_STRIP/2;
+    var waveWidth = 10;
+    var wavePosition = 0;
 
 
     timer = setInterval(function() {
         // create an array to hold the data for all the strips at once
         // loop
         var strips = [];
-        for (var i = 0; i < NUM_STRIPS; i ++){
-            var stripId = i;
+        for (var stripId = 0; stripId< NUM_STRIPS; stripId ++){
             var s = new PixelStrip(stripId,PIXELS_PER_STRIP);
 
-            // this sets random pixels on for each strip
-            // it also does a different color for wach of the 4 strips
-            /*
-            if (i == 0 || i == 4){
-                s.getRandomPixel().setColor(255,0,0, 0.1);
-            } else if (i == 1 || i == 5){
-                s.getRandomPixel().setColor(0,255,0, 0.1);
-            } else if (i == 2 || i == 6){
-                s.getRandomPixel().setColor(0,0,255, 0.1);
-            } else if (i == 3 || i == 7){
-                s.getRandomPixel().setColor(100,0,55, 0.1);
+            var startIdx = waveHeight+wavePosition;
+            for (var i = startIdx, j = waveWidth; i <PIXELS_PER_STRIP &&  i > waveHeight && j > 0; i --, j--){
+                // right wave
+                var p = s.getPixel(i);
+                p.setColor(255,0,0,(j/waveWidth));
             }
-            */
 
-            //s.getRandomPixel().setColor(255,0,0, 0.1);
-            // set the whole strip blue
-            s.setStripColor(0,255,255, 0.01);
+            var startIdx = waveHeight-wavePosition;
+            for (var i = startIdx, j = waveWidth; i > 0 &&  i < waveHeight && j > 0; i ++, j--){
+                // right wave
+                var p = s.getPixel(i);
+                p.setColor(255,0,0,(j/waveWidth));
+            }
 
+            // to show combined systems also set a random pixel blue
+            // set a random pixel blue
+            s.getRandomPixel().setColor(0,0,255, 0.1);
             // render the strip data into the correct format for sending
             // to the pixel pusher controller
             var renderedStripData = s.getStripData();
             // add this data to our list of strip data to send
             strips.push(renderedStripData);
         }
-
-        //console.log(strips.length)
-
         // inform the controller of the new strip frame
         controller.refresh(strips);
 
+        wavePosition = (wavePosition + 1) % waveHeight;
+
     }, UPDATE_FREQUENCY_MILLIS);
+
 }).on('error', function(err) {
   console.log('PixelPusher Error: ' + err.message);
 });
